Remove element place
OK.

So we each replace the three and the three and we need to return basically the new length which is gonna

be just remaining to do so.

Let's take a look at another example.

Basically 0 1 2 2 3 0 4 2.

Now we are given to in this case we replace to remove basically two from all the instances and then

return the new length which is going to be five.